// frontend/src/app/api/model/create-person-request.ts
/**
 * Coach Buddy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CreatePhoneNumberRequest } from './create-phone-number-request';


export interface CreatePersonRequest { 
    firstName: string;
    lastName: string;
    birthDate?: string;
    nickname?: string;
    notes?: string;
    phoneNumbers?: Array<CreatePhoneNumberRequest>;
}